lesBonComptes-Bilel/requirements.txt
appier==1.32.0
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
dnspython==2.5.0
email-validator==2.1.0.post1
filelock==3.13.1
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Session==0.4.0
Flask-WTF==1.2.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mongoengine==0.27.0
numpy==1.26.4
pandas==2.2.0
paypalrestsdk==1.13.3
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.1
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
setuptools==68.2.0
six==1.16.0
stripe==2.65.0
stripe-api==0.2.1
tldextract==5.1.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
Werkzeug==3.0.1
WTForms==3.1.2
zope.event==5.0
zope.interface==6.2
flask-mongoengine-3==1.0.8
Authlib == 1.3.0
oauth==1.0.1
flask-dance==7.1.0
flask-cors == 4.0.0